refactor simple example
